Improved implementation of UMap to use UAsset base code
